// src/app/feature/shared/models/company.ts
export class Company {
  CompanyID: number;
  CompanyCode: string;
  CompanyName: string;
  CompanyEmail: string;
  CompanyAddress: string;
  CompanyWebsite: string;
  CompanyDBSchema: string;
  CompanyDBName: string;
  CompanyUsersDomain: string;
  PhoneNumber: string;
  SolutionTypeID: string;
  CompanyDBServerID: number;
  UserEmail: string;
  UserName: string;
  UserPassword: string;
  IsDeleted: boolean;
  PageNumber: number;
  PageSize: number;
  LogoURL: string;
}